상세 화면 ViewModel 생성 및 `NetworkFetcher` 구조 개선
*   **상세 화면**: `DetailScreen`, `DetailViewModel`, `DetailUseCase`를 추가하여 영화 및 TV 상세 정보 표시 기능 구현. `HiltViewModel`의 `AssistedFactory`를 사용하여 `IMediaModel`을 ViewModel에 주입.
*   **UseCase**: `DetailUseCase`를 추가하여 미디어 타입(영화/TV)에 따라 적절한 Repository의 상세 정보 API를 호출하도록 분기 처리.
*   **UI 및 Preview**: `DetailScreen`에 `TopAppBar`를 추가하고, `NetworkResult`의 상태(캐시 데이터, 로딩 완료)에 따른 프리뷰를 `createTestInstance`를 통해 구현.
*   **Test Data**: `MovieDetailModel` 및 `TVDetailModel`에 Preview와 테스트를 위한 `createTestInstance` 팩토리 메서드 추가.
*   **Navigation**: `MovieApp`에서 `Detail` Route 진입 시 `hiltViewModel`을 통해 `DetailViewModel`을 생성하고 `DetailScreen`에 전달하도록 설정.
*   **NetworkFetcher**: `NetworkResult`의 생성자 파라미터 순서를 변경하여 `data`를 앞으로 이동시키고, `hasData` 함수를 `inline`으로 변경.

+++ data/src/main/java/com/gondev/data/repository/MovieRepositoryImpl.kt
@@ -19,6 +19,6 @@ class MovieRepositoryImpl @Inject constructor(
     override suspend fun getSearch(query: String) =
         apiService.getSearchMovie(query).toDomain { it.toDomain() }
 
-    override suspend fun getDetail(mediaId: Long) =
+    override suspend fun getDetail(mediaId: Int) =
         apiService.getMovieDetail(mediaId).toDomain()
 }
